feat: simplify badge/popover interaction
